requirements.txt
anyio==3.7.1
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.1.0
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
executing==2.0.0
gunicorn==21.2.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.16.1
jedi==0.19.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
matplotlib-inline==0.1.6
packaging==23.2
pacote==0.3.0
parso==0.8.3
pickleshare==0.7.5
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.5.2
pytest-testdox==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rest-framework-simplejwt==0.0.2
rfc3986==1.5.0
rpds-py==0.10.6
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.12.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
wcwidth==0.2.8
whitenoise==6.6.0
